requirements.txt
# Core Deep Learning
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2

# YOLOv8 and related
ultralytics>=8.0.196
supervision>=0.18.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# XML Processing (for Pascal VOC)
lxml>=4.9.0
xmltodict>=0.13.0

# Progress bars and utilities
tqdm>=4.65.0
PyYAML>=6.0
albumentations>=1.3.0

# For ZED camera (optional, install later if needed)
# zed-python-api==3.8

# Jupyter for experimentation
jupyter
notebook
ipywidgets